# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MarketStock {
  symbol: string;
  sector: string;
  ltp: number;
  ycp: number;
  change: number;
  high: number;
  low: number;
  volume: number;
  updatedAt: string;
}

export interface HistoryPoint {
  time: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface IngestionStatus {
  total: number;
  completed: number;
  pending: number;
  failed: number;
  inProgress: string;
  nextRunAt: string;
  lastSynced: string;
  lastRecords: number;
}

export interface MarketOverview {
  market: string;
  asOf: string;
  stocks: MarketStock[];
  sectors: string[];
  ingestion: IngestionStatus;
}

export interface IngestionAction {
  accepted: boolean;
  message: string;
}

export interface IngestionTrackerRow {
  symbol: string;
  status: string;
  totalRecordsInserted: number;
  lastUpdated: string;
}

export interface IngestionLog {
  time: string;
  symbol: string;
  status: string;
  message: string;
  records: number;
}

export type GetMarketStocksParams = {
sector?: string;
search?: string;
};