# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market/stocks:
    get:
      operationId: getMarketStocks
      tags: [market]
      summary: List current DSE market stocks
      parameters:
        - name: sector
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Current stock list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketStock"
  /market/stocks/{symbol}/history:
    get:
      operationId: getStockHistory
      tags: [market]
      summary: Get historical OHLCV data for a stock
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryPoint"
...
components:
  schemas:
...
    MarketStock:
      type: object
      properties:
        symbol:
          type: string
        sector:
          type: string
        ltp:
          type: number
        ycp:
          type: number
        change:
          type: number
        high:
          type: number
        low:
          type: number
        volume:
          type: integer
        updatedAt:
          type: string
      required: [symbol, sector, ltp, ycp, change, high, low, volume, updatedAt]
    HistoryPoint:
      type: object
      properties:
        time:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: integer
      required: [time, open, high, low, close, volume]
...